# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Chat APP API",
        "description": "API documentation for Chat Application - A social media chat application with posts, comments, likes, friendships, and real-time messaging",
        "contact": {
            "email": "support@chatapp.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://my-default-host.com",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/users/search": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Search users",
                "description": "Search for users by name or email",
                "operationId": "20994571ae6f2f07d415b885f7c28471",
                "parameters": [
                    {
                        "name": "q",
                        "in": "query",
                        "description": "Search query (name or email)",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "john"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of users matching the search query",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "name": {
                                                "type": "string",
                                                "example": "John Doe"
                                            },
                                            "email": {
                                                "type": "string",
                                                "format": "email",
                                                "example": "john@example.com"
                                            },
                                            "profile_image": {
                                                "type": "string",
                                                "nullable": true
                                            },
                                            "last_seen_at": {
                                                "type": "string",
                                                "format": "date-time",
                                                "nullable": true
                                            },
                                            "friendship_id": {
                                                "type": "integer",
                                                "nullable": true
                                            },
                                            "friendship_status": {
                                                "type": "string",
                                                "enum": [
                                                    "friends",
                                                    "pending_outgoing",
                                                    "pending_incoming"
                                                ],
                                                "nullable": true
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Authentication using Laravel session-based authentication",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}